Removed class SquaredAbs and added intrinsic usage
Intrinsics are now used in the implementation of the new methods
operator>() and lteToPixels() in the SimdUnion classes.
In addition a setValue method on setting values in registers directly
has been added. The old setValue() method has been renamed to
setValueInReg().
Method setRealValuesReverseInReg() has been added. This method is needed
because of the new lteToPixels() methods.